+++ data/admin/microorganism.py
@@ -3,9 +3,13 @@ from django.contrib import admin
 
 from simple_history.admin import SimpleHistoryAdmin
 
-from data.models import Declaration, Microorganism
+from data.models import Microorganism
 
-from .abstract_admin import ChangeReasonAdminMixin, ChangeReasonFormMixin
+from .abstract_admin import (
+    ChangeReasonAdminMixin,
+    ChangeReasonFormMixin,
+    RecomputeDeclarationArticleAtIngredientSaveMixin,
+)
 
 
 class MicroorganismForm(ChangeReasonFormMixin):
@@ -18,7 +22,8 @@ class MicroorganismForm(ChangeReasonFormMixin):
 
 
 @admin.register(Microorganism)
-class MicroorganismAdmin(ChangeReasonAdminMixin, SimpleHistoryAdmin):
+class MicroorganismAdmin(RecomputeDeclarationArticleAtIngredientSaveMixin, ChangeReasonAdminMixin, SimpleHistoryAdmin):
+    declaredingredient_set = "declaredmicroorganism_set"
     form = MicroorganismForm
     fieldsets = [
         (
@@ -76,20 +81,3 @@ class MicroorganismAdmin(ChangeReasonAdminMixin, SimpleHistoryAdmin):
         "siccrf_species",
     )
     search_fields = ["id", "name"]
-
-    def save_model(self, request, obj, form, change):
-        super().save_model(request, obj, form, change)
-        # recalcul de l'article pour les déclarations concernées
-        if change and form["is_risky"].has_changed():
-            for declaration in Declaration.objects.filter(
-                id__in=obj.declaredmicroorganism_set.values_list("declaration_id", flat=True),
-                status__in=(
-                    Declaration.DeclarationStatus.AWAITING_INSTRUCTION,
-                    Declaration.DeclarationStatus.ONGOING_INSTRUCTION,
-                    Declaration.DeclarationStatus.AWAITING_VISA,
-                    Declaration.DeclarationStatus.OBSERVATION,
-                    Declaration.DeclarationStatus.OBJECTION,
-                ),
-            ):
-                declaration.assign_calculated_article()
-                declaration.save()
