+++ data/admin/plant.py
@@ -4,9 +4,13 @@ from django.db import models
 
 from simple_history.admin import SimpleHistoryAdmin
 
-from data.models import Declaration, Plant, PlantSynonym
+from data.models import Plant, PlantSynonym
 
-from .abstract_admin import ChangeReasonAdminMixin, ChangeReasonFormMixin
+from .abstract_admin import (
+    ChangeReasonAdminMixin,
+    ChangeReasonFormMixin,
+    RecomputeDeclarationArticleAtIngredientSaveMixin,
+)
 
 
 class PlantSynonymInline(admin.TabularInline):
@@ -39,7 +43,8 @@ class PlantForm(ChangeReasonFormMixin):
 
 
 @admin.register(Plant)
-class PlantAdmin(ChangeReasonAdminMixin, SimpleHistoryAdmin):
+class PlantAdmin(RecomputeDeclarationArticleAtIngredientSaveMixin, ChangeReasonAdminMixin, SimpleHistoryAdmin):
+    declaredingredient_set = "declaredplant_set"
     form = PlantForm
     fieldsets = [
         (
@@ -99,20 +104,3 @@ class PlantAdmin(ChangeReasonAdminMixin, SimpleHistoryAdmin):
         "siccrf_family",
     )
     search_fields = ["id", "name"]
-
-    def save_model(self, request, obj, form, change):
-        super().save_model(request, obj, form, change)
-        # recalcul de l'article pour les déclarations concernées
-        if change and form["is_risky"].has_changed():
-            for declaration in Declaration.objects.filter(
-                id__in=obj.declaredplant_set.values_list("declaration_id", flat=True),
-                status__in=(
-                    Declaration.DeclarationStatus.AWAITING_INSTRUCTION,
-                    Declaration.DeclarationStatus.ONGOING_INSTRUCTION,
-                    Declaration.DeclarationStatus.AWAITING_VISA,
-                    Declaration.DeclarationStatus.OBSERVATION,
-                    Declaration.DeclarationStatus.OBJECTION,
-                ),
-            ):
-                declaration.assign_calculated_article()
-                declaration.save()
